requirements.test.txt
astroid==3.3.10
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
coverage==7.9.1
dill==0.4.0
exceptiongroup==1.3.0
Flask==3.1.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
pip==25.1
platformdirs==4.3.8
pluggy==1.6.0
pydub==0.25.1
Pygments==2.19.2
pylint==3.3.7
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.1.1
pywgb==1.0.4
requests==2.32.4
setuptools==78.1.1
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
